pyproject.toml
# Installation:
# - For local development (macOS/no GPU): pip install -e ".[minimal]"
# - For Linux server with GPU: pip install -e ".[linux-gpu]" --extra-index-url https://pypi.nvidia.com
# - For development tools: pip install -e ".[dev]"
# Note: On Linux, PyTorch CUDA support should be installed separately with:
#       pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "CourtPressGER"
version = "0.1.0"
description = "German Court Press Release Dataset"
readme = "README.md"
requires-python = ">=3.8"
authors = [{name = "Sebastian Nagl", email = "sebastian.nagl@tum.de"}]
license = {text = "MIT"}
dependencies = [
    # Core Data Science Stack
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    
    # NLP & Transformer Tools (CPU versions)
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.0",
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    "gensim>=4.3.0",
    "bertopic>=0.15.0",
    "keybert>=0.7.0",
    "flair>=0.12.0",
    "fasttext-wheel>=0.9.2",
    
    # Clustering & Dimensionality Reduction
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.4",
    
    # Visualization
    "matplotlib-venn>=0.11.9",
    
    # Jupyter Environment
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
    "tqdm>=4.66.0",
    
    # File Handling
    "pathlib",
    "datasets>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.12.0",
]

# GPU dependencies for Linux server
linux-gpu = [
    # RAPIDS Ecosystem (GPU-accelerated Data Science)
    "cudf-cu12>=24.2.0",
    "cuml-cu12>=24.2.0",
    "cupy-cuda12x>=13.0.0",
    "rmm-cu12>=24.2.0",
    "cugraph-cu12>=24.2.0",
    "cuspatial-cu12>=24.2.0",
    "dask-cuda>=24.2.0",
    
    # GPU-accelerated Deep Learning
    "torch>=2.1.0",  # CUDA version installed separately
    "torchtext>=0.16.0",
    "torch-scatter>=2.1.0",
    "torch-sparse>=0.6.0",
    
    # GPU-accelerated NLP
    "spacy-cuda12x>=3.7.0",
    "spacy-cupy>=3.7.0",
    
    # Additional GPU Utilities
    "numba>=0.58.0",
    "nvtx>=0.2.5",
]

# Minimal dependencies for local development
minimal = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "spacy>=3.7.0",
    "jupyter>=1.0.0",
    "tqdm>=4.66.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.setuptools]
packages = []
include-package-data = false

[tool.setuptools.exclude-package-data]
"*" = ["data/raw/*", "data/processed/*", "data/interim/*"]

[project.scripts]
courtpress-analyze = "courtpress.scripts.run_analysis:main"
courtpress-clean = "courtpress.scripts.run_cleaning:main"
courtpress-prompts = "courtpress.scripts.generate_prompts:main"
courtpress-combined = "courtpress.scripts.run_combined_analysis:main"
